django_backend/requirements.txt
django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
gunicorn>=21.2.0
django-filter>=23.5
djangorestframework-simplejwt>=5.3.1
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
psutil>=5.9.0
